Mellanox system drivers backport to kernel 4.9
This series of patches introduces support for Mellanox systems.
Most of the patches is a backporting from the kernels 4.10  4.14
for Mellanox system drivers that have been applied upstream and
a few patches contain current Mellanox upstream work  patches
which are in review process for acceptances for kernels 4.16  4.17.
This series introduces Mellanox i2c controller, mux, platform,
hotplug, led, asic drivers, TI voltage controller driver and the
extension of Mellanox kernel config.

In series file all the patches are listed and all old patches, relevant
for old 3.16 kernel are commented out.

Signed-off-by: Vadim Pasternak <vadimp@melanox.com>